requirements.txt
asyncio
websockets
psycopg2
python-dotenv